apps/api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crum==0.7.9
django-environ==0.9.0
django-import-export==3.0.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-shopping-cart==0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpie==3.2.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.3
Pillow==9.3.0
psycopg2==2.9.5
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tablib==3.3.0
tropipy==0.0.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
